[TAN-8009] Simplify
